requirements.txt
# FastAPI và server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Google Gemini API
google-generativeai>=0.3.0

# Machine Learning và NLP
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Utilities
numpy>=1.24.0
pydantic>=2.0.0

# Firebase Admin SDK
firebase-admin>=6.0.0
google-auth>=2.23.0,<2.42.0  # Tương thích với google-auth-oauthlib
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

